refactor : user 도메인 분리
