+++ blackWhiteDeliveryDriver/src/main/java/com/sparta/blackwhitedeliverydriver/user/dto/SignupRequestDto.java
@@ -1,7 +1,6 @@
-package com.sparta.blackwhitedeliverydriver.dto;
+package com.sparta.blackwhitedeliverydriver.user.dto;
 
-import com.sparta.blackwhitedeliverydriver.entity.UserRoleEnum;
-import jakarta.validation.constraints.Email;
+import com.sparta.blackwhitedeliverydriver.user.entity.UserRoleEnum;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Null;
@@ -9,7 +8,6 @@ import jakarta.validation.constraints.Pattern;
 import jakarta.validation.constraints.Size;
 import lombok.Builder;
 import lombok.Getter;
-import lombok.Setter;
 
 @Getter
 @Builder
