+++ blackWhiteDeliveryDriver/src/main/java/com/sparta/blackwhitedeliverydriver/user/entity/User.java
@@ -1,7 +1,9 @@
-package com.sparta.blackwhitedeliverydriver.entity;
+package com.sparta.blackwhitedeliverydriver.user.entity;
 
-import com.sparta.blackwhitedeliverydriver.dto.SignupRequestDto;
-import com.sparta.blackwhitedeliverydriver.dto.UpdateUserRequestDto;
+import com.sparta.blackwhitedeliverydriver.address.entity.Address;
+import com.sparta.blackwhitedeliverydriver.common.entity.BaseEntity;
+import com.sparta.blackwhitedeliverydriver.user.dto.SignupRequestDto;
+import com.sparta.blackwhitedeliverydriver.user.dto.UpdateUserRequestDto;
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
