+++ blackWhiteDeliveryDriver/src/main/java/com/sparta/blackwhitedeliverydriver/user/exception/ExceptionMessage.java
@@ -1,4 +1,4 @@
-package com.sparta.blackwhitedeliverydriver.exception;
+package com.sparta.blackwhitedeliverydriver.user.exception;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
